+++ out/test/test.js
@@ -1,19 +1,20 @@
-export const Test1 = `{
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.Test3 = exports.Test2 = exports.Test1 = void 0;
+exports.Test1 = `{
   a: string;
   b: number;
   c: [];
   d: undefined;
 }`;
-
-export const Test2 = `{
+exports.Test2 = `{
   a: string; // a is something good. Magna do ad est aliqua anim sint nulla occaecat. Consectetur dolor reprehenderit ea mollit ipsum amet.Consectetur non ad duis Lorem in nostrud qui fugiat ea est laboris culpa cupidatat.
   b: number; // Est dolore ea ex enim aliqua deserunt anim occaecat.
   // Dolor proident pariatur mollit duis irure.
   c: [];
   d: undefined;
 };`;
-
-export const Test3 = `{
+exports.Test3 = `{
   adsf: string;
   b: number;
   cvcxv: [];
@@ -33,3 +34,4 @@ export const Test3 = `{
   k: Object;
   o: { t: { a: string; b: []; c: { h: { j: string; n: number } } } };
 };`;
+//# sourceMappingURL=test.js.map